Avoided ASCII-cast warnings for WebKit.
Normally they won't be enabled anyway, but if you build webkit from
within the Qt qt5 repository it will pick up Qt's default build
settings, which do enable it. We need to disable them because
warnings are treated as errors and there are way too many of them in
the WebKit code.