Removed modular references after support for the flag was removed.
Support was removed because the Qt Modularization project decided we
don't need it. It's better to base decisions on the available
information, as demonstrated in the patch by checking
QT.phonon.includes.

In addition, remove the reference to uitools. It has been moved to
QtBase now, and therefore is always available.