Fix build on linux-i386 where gcc would produce warnings
It is the intention to _not_ use -Werror for production builds. Unfortunately
the qmake expression had an error that made the -Werror suppression only work
on x86-64 but not on i386. A surrounding if() as suggested by Ossi fixes this.

In trunk this is already fixed differently by the use of an extra scope.

Task: QTBUG-23987